- Added a duplicate .png for missing resource (bosstexture) - Edited a corrupt tag within MapEditor.csproj - Fixed a merge conflict within MapEditor.csproj.FileListAbsolute.txt
Not sure what initially caused this issue, but maybe it's just github for windows?  I'll throw out a recommendation for SourceTree with future projects.
Also, bear in mind that this might be a reverted version.  You should check to make sure your code is up to date with your most recent changes during class.
Also, after the class is done, you should maybe add a license and open source this and your other engine?  Open Sourcing looks awesome on resumes.

Signed-off-by: Daniel Shumway <shumway.danny@gmail.com>